requirements.txt
asgiref==3.7.2
black==23.3.0
click==8.1.3
Django==4.2.1
django-ninja==0.22.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.1
pydantic==1.10.8
sqlparse==0.4.4
typing_extensions==4.6.2
